Move factory tests to own module
